Splits CLIViewTest into sections
