+++ src/test/java/mike/cli/CLIViewTest.java
@@ -16,15 +16,11 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import mike.datastructures.Entity.visibility;
 import mike.datastructures.Model;
 import mike.HelperMethods;
 import mike.controller.CLIController;
-import mike.controller.Controller;
-import mike.datastructures.Entity;
-import mike.datastructures.Method;
-import mike.datastructures.Relationship.Type;
 
-import mike.view.CLIView;
 import mike.view.ViewTemplate;
 
 public class CLIViewTest {
@@ -61,1414 +57,6 @@ public class CLIViewTest {
         err.reset();
     }
 	
-    //**********************************************************************************************//
-    //**                                CONSTRUCTOR TEST                                          **//
-    //**                                (INITIALIZATION)                                          **//
-    //**********************************************************************************************//
-	
-    @Test
-    public void initTest(){
-	assertTrue("CLI is null; Should not be null.", control != null);
-	assertTrue("CLI's model is not empty; Should be empty.", control.getModel().empty());
-    }
-    
-    //**********************************************************************************************//
-    //**                                CLASS TESTS                                               **//
-    //**                         (CREATE, RENAME, DELETE)                                         **//
-    //**********************************************************************************************//
-	
-    @Test
-    public void createClassTest() {
-	// Test creating one class
-	String[] commands = {"create", "class", "c1"};
-	control.evaluateCommand(commands);
-	assertTrue("CLI did not make a class; Should be one class called c1.", model.containsEntity("c1"));
-	assertEquals("CLI made more or less than one class", 1, model.getEntities().size());
-	
-	// Test creating 4 more classes
-	commands[2] = "c2";
-	control.evaluateCommand(commands);
-	commands[2] = "c3";
-	control.evaluateCommand(commands);
-	commands[2] = "c4";
-	control.evaluateCommand(commands);
-	commands[2] = "c5";
-	control.evaluateCommand(commands);
-	
-	assertTrue("CLI did not make a class; Should be class called c2.", model.containsEntity("c2"));
-	assertTrue("CLI did not make a class; Should be class called c3.", model.containsEntity("c3"));
-	assertTrue("CLI did not make a class; Should be class called c4.", model.containsEntity("c4"));
-	assertTrue("CLI did not make a class; Should be class called c5.", model.containsEntity("c5"));
-	assertEquals("CLI did not make classes; Should be 5 classes.", 5, model.getEntities().size());
-	
-	// Test NOT creating with invalid length
-        String[] invalidLengthCommand = {"create", "class", "badLength", "Yes."};
-        control.evaluateCommand(invalidLengthCommand);
+   
 
-        assertFalse("CLI created a class with bad length; Class named 'badLength' should not exist.", model.containsEntity("badLength"));
-        assertEquals("CLI made a class; Should be 5 classes.", 5, model.getEntities().size());
-
-        // Test NOT making an EXISTING class
-        commands[2] = "c1";
-        control.evaluateCommand(commands);
-
-        assertEquals("CLI created a class that already existed; Length should be 5.", 5, model.getEntities().size());
-    }
-    
-    @Test
-    public void renameClassTest() {
-        // Test setup of class creation
-        String[] commands = {"create", "class", "c1"};
-        control.evaluateCommand(commands);
-        commands[2] = "c2";
-        control.evaluateCommand(commands);
-
-        assertTrue("CLI did not make a class; Should be class called c1.", model.containsEntity("c1"));
-        assertTrue("CLI did not make a class; Should be class called c2.", model.containsEntity("c2"));
-
-        // Test rename a class
-        String[] rename = {"rename", "class", "c1", "c3"};
-        control.evaluateCommand(rename);
-
-        assertTrue("CLI did not rename a class; Should be class called c3.", model.containsEntity("c3"));
-        assertFalse("CLI did not remove/rename old class; c1 should no longer exist.", model.containsEntity("c1"));
-
-        // Test NOT renaming with invalid length
-        String[] invalidLengthCommand = {"create", "class", "c3", "c1", "badLength"};
-        control.evaluateCommand(invalidLengthCommand);
-
-        assertFalse("CLI created a class with bad length; Class named 'c1' should not exist.", model.containsEntity("c1"));
-        assertEquals("CLI made a class; Should be 2 classes.", 2, model.getEntities().size());
-
-
-        // Test rename a NONEXISTENT class
-        rename[2] = "c4";
-        rename[3] = "c5";
-        control.evaluateCommand(rename);
-
-        assertTrue("CLI renamed a class. c2 should exist.", model.containsEntity("c2"));
-        assertTrue("CLI renamed a class. c3 should exist.", model.containsEntity("c3"));
-        assertFalse("CLI created a new class. c4 should not exist.", model.containsEntity("c4"));
-        assertFalse("CLI created a new class. c5 should not exist.", model.containsEntity("c5"));
-
-        // Test rename an EXISTING class to another EXISTING class
-        rename[2] = "c3";
-        rename[3] = "c2";
-        control.evaluateCommand(rename);
-
-        assertTrue("CLI renamed a class. c2 should exist.", model.containsEntity("c2"));
-        assertTrue("CLI renamed a class. c3 should exist.", model.containsEntity("c3"));
-        assertEquals("CLI does not have 2 classes. Model should have two entities.", 2, model.getEntities().size());
-    }
-    
-    @Test
-    public void deleteClassTest() {
-        // Test setup of class creation
-        String[] createClass = {"create", "class", "c1"};
-        control.evaluateCommand(createClass);
-        createClass[2] = "c2";
-        control.evaluateCommand(createClass);
-
-        assertTrue("CLI did not make a class; Should be class called c1.", model.containsEntity("c1"));
-        assertTrue("CLI did not make a class; Should be class called c2.", model.containsEntity("c2"));
-
-        // Test deleting a class
-        String[] delete = {"delete", "class", "c1"};
-        control.evaluateCommand(delete);
-
-        assertFalse("CLI did not delete a class; c1 should not exist.", model.containsEntity("c1"));
-        assertTrue("CLI did something to another class; c2 should be okay.", model.containsEntity("c2"));
-
-        // Test NOT deleting with invalid length
-        String[] invalidLengthCommand = {"delete", "class", "c2", "badLength"};
-        control.evaluateCommand(invalidLengthCommand);
-
-        assertTrue("CLI deleted a class with bad length; Class named 'c2' should exist.", model.containsEntity("c2"));
-
-        // Test NOT deleting a NONEXISTENT class
-        control.evaluateCommand(delete);
-        
-        assertFalse("CLI did not do something weird. c1 should not exist.", model.containsEntity("c1"));
-        assertTrue("CLI deleted a class it wasn't supposed to; Class 'c2' should exist.", model.containsEntity("c2"));
-
-        // Test deleting classes that have fields
-        createClass[2] = "c1";
-        control.evaluateCommand(createClass);
-        String[] createField = {"create", "field", "c1", "public", "int", "f1"};
-        control.evaluateCommand(createField);
-        control.evaluateCommand(delete);
-        
-        assertFalse("CLI did not delete the class; c1 should not exist.", model.containsEntity("c1"));
-        
-        // Test deleting classes that have methods
-        control.evaluateCommand(createClass);
-        String[] createMethod = {"create", "method", "c1", "public", "int", "m1"};
-        control.evaluateCommand(createMethod);
-        control.evaluateCommand(delete);
-        
-        assertFalse("CLI did not delete the class. c1 should not exist.", model.containsEntity("c1"));
-        
-        // Test deleting classes with methods and parameters
-        control.evaluateCommand(createClass);
-        String[] createParameter = {"create", "parameter", "c1", "m1", "int", "p1"};
-        control.evaluateCommand(createMethod);
-        control.evaluateCommand(createParameter);
-        control.evaluateCommand(delete);
-        
-        assertFalse("CLI did not delete the class. c1 should not exist.", model.containsEntity("c1"));
-
-        // Test deleting class with a relationship
-        control.evaluateCommand(createClass);
-        String[] createRelationship = {"create", "relationship", "realization", "c1", "c2"};
-        control.evaluateCommand(createRelationship);
-        
-        assertTrue("CLI did not create the relationship. Should be a relationship from c1 to c2.", model.containsRelationship(Type.REALIZATION, "c1", "c2"));
-        
-        control.evaluateCommand(delete);
-        
-        assertFalse("CLI did not delete the class. c1 should not exist.", model.containsEntity("c1"));
-        assertFalse("CLI did not delete the relationship. Realization from c1 to c2 should not exist.", model.containsRelationship(Type.REALIZATION, "c1", "c2"));
-    }
- 
-    //**********************************************************************************************//
-    //**                                FIELD TESTS                                               **//
-    //**                         (CREATE, RENAME, DELETE)                                         **//
-    //**********************************************************************************************//
-    
-    @Test
-    public void createFieldTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	
-	// Create one field
-	String[] fieldCommands = {"create", "field", "c1", "public", "int", "f1"};
-	control.evaluateCommand(fieldCommands);
-	assertTrue("CLI did not make a field; Should be one field called f1.", c1.containsField("f1"));
-	assertEquals("c1 made more or less than one field.", 1, c1.getFields().size());
-	
-	// Create multiple fields
-	fieldCommands[5] = "f2";
-	control.evaluateCommand(fieldCommands);
-	fieldCommands[5] = "f3";
-	control.evaluateCommand(fieldCommands);
-	fieldCommands[5] = "f4";
-	control.evaluateCommand(fieldCommands);
-	assertTrue("CLI did not make a field; Should be one field called f2.", c1.containsField("f2"));
-	assertTrue("CLI did not make a field; Should be one field called f3.", c1.containsField("f3"));
-	assertTrue("CLI did not make a field; Should be one field called f4.", c1.containsField("f4"));
-	assertEquals("c1 has more or less than 4 fields", 4, c1.getFields().size());
-	
-	// Create field with wrong length of commands
-	String[] invalidLengthCommand = {"create", "field", "c1", "public", "int", "tooLong", "badLength"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertFalse("CLI made a field; Shouldn't have since too many arguments.", c1.containsField("badLength"));
-	assertFalse("CLI made a field; Shouldn't have since too many arguments.", c1.containsField("tooLong"));
-	assertEquals("CLI made a field; Shouldn't have since too many arguments.", 4, c1.getFields().size());
-	
-	// Create field with non-existing class
-	fieldCommands[2] = "c3";
-	fieldCommands[5] = "f6";
-	control.evaluateCommand(fieldCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertFalse("CLI made a field in a class that doesn't exist; Should not have", c1.containsField("f6"));
-	assertEquals("c1 has more or less than 4 fields.", 4, c1.getFields().size());
-	
-	// Create field with same name
-	fieldCommands[5] = "f1";
-	control.evaluateCommand(fieldCommands);
-	assertEquals("CLI made a field; Field name already existed.", 4, c1.getFields().size());
-	
-	// Create field in separate classes
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	fieldCommands[2] = "c2";
-	fieldCommands[5] = "f5";
-	control.evaluateCommand(fieldCommands);
-	assertTrue("CLI did not make a field; Should be one field called f5.", c2.containsField("f5"));
-	assertEquals("c2 has more or less than one field.", 1, c2.getFields().size());
-	
-	// Can create the same field name in another class
-	fieldCommands[5] = "f1";
-	control.evaluateCommand(fieldCommands);
-	assertTrue("CLI did not make a field; Should be one field called f1.", c2.containsField("f1"));
-	assertTrue("CLI deleted field f5; f5 should still exist.", c2.containsField("f5"));
-	assertEquals("c2 has more or less than two fields.", 2, c2.getFields().size());
-
-	// Can not create a field with same name, but different type
-	fieldCommands[4] = "int";
-	fieldCommands[5] = "f5";
-	control.evaluateCommand(fieldCommands);
-	assertTrue("CLI broke", c2.containsField("f5"));
-	assertEquals("c2 has more or less than two fields.", 2, c2.getFields().size());
-    }
-
-    @Test
-    public void renameFieldTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] createFieldCommands = {"create", "field", "c1", "public", "int", "f1"};
-	control.evaluateCommand(createFieldCommands);
-	
-	// Rename one field
-	String[] renameFieldCommands = {"rename", "field", "c1", "f1", "newf1"};
-	control.evaluateCommand(renameFieldCommands);
-	assertFalse("c1 still has a f1; Should not have any more f1.", c1.containsField("f1"));
-	assertTrue("c1 still has a f1; Should not have any more f1.", c1.containsField("newf1"));
-	assertEquals("c1 has more or less than one fields.", 1, c1.getFields().size());
-	
-	// Rename multiple fields
-	createFieldCommands[5] = "f2";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f3";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f4";
-	control.evaluateCommand(createFieldCommands);
-	renameFieldCommands[3] = "f4";
-	renameFieldCommands[4] = "newf4";
-	control.evaluateCommand(renameFieldCommands);
-	renameFieldCommands[3] = "f3";
-	renameFieldCommands[4] = "newf3";
-	control.evaluateCommand(renameFieldCommands);
-	renameFieldCommands[3] = "f2";
-	renameFieldCommands[4] = "newf2";
-	control.evaluateCommand(renameFieldCommands);
-	assertFalse("c1 still has field f2; Should not have it.", c1.containsField("f2"));
-	assertFalse("c1 still has field f3; Should not have it.", c1.containsField("f3"));
-	assertFalse("c1 still has field f4; Should not have it.", c1.containsField("f4"));
-	assertTrue("c1 doesn't have field newf2; Should have it.", c1.containsField("newf2"));
-	assertTrue("c1 doesn't have field newf3; Should have it.", c1.containsField("newf3"));
-	assertTrue("c1 doesn't have field newf4; Should have it.", c1.containsField("newf4"));
-	assertEquals("c1 has more or less than four fields.", 4, c1.getFields().size());
-	
-	// Rename field with wrong length of commands
-	String[] invalidLengthCommand = {"rename", "field", "c1", "newf1", "tooLong", "reallyLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("c1 no longer has newf1; Should still have it after an error.", c1.containsField("newf1"));
-	assertFalse("c1 renamed field newf1; Shouldn't have since parameters are invalid.", c1.containsField("tooLong"));
-	assertFalse("c1 renamed field newf1; Shouldn't have since parameters are invalid.", c1.containsField("reallyLong"));
-	assertEquals("c1 has more or less than four fields.", 4, c1.getFields().size());
-	
-	// Rename field with non-existing class
-	renameFieldCommands[2] = "c3";
-	renameFieldCommands[3] = "newf1";
-	renameFieldCommands[4] = "wrongf1";
-	control.evaluateCommand(renameFieldCommands);
-	assertTrue("c1 no longer has newf1; Should still have it after an error.", c1.containsField("newf1"));
-	assertFalse("c1 contains wrongf1; Shouldn't have since class c3 doesn't exist.", c1.containsField("wrongf1"));
-	assertEquals("CLI has more or less than one classes.", 1, model.getEntities().size());
-	assertEquals("c1 has more or less than four fields.", 4, c1.getFields().size());
-	
-	// Rename field with non-existing field
-	renameFieldCommands[2] = "c1";
-	renameFieldCommands[3] = "f5";
-	control.evaluateCommand(renameFieldCommands);
-	assertFalse("CLI broke; Should not have", c1.containsField("f5"));
-	assertFalse("c1 contains wrongf1; Should not since f5 doesn't exist", c1.containsField("wrongf1"));
-	assertEquals("c1 has more or less than 4 fields.", 4, c1.getFields().size());	
-
-	// Rename field in separate classes (with different or same name fields)
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	createFieldCommands[2] = "c2";
-	createFieldCommands[5] = "newf1";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f5";
-	control.evaluateCommand(createFieldCommands);
-	renameFieldCommands[2] = "c2";
-	renameFieldCommands[3] = "f5";
-	renameFieldCommands[4] = "newf5";
-	control.evaluateCommand(renameFieldCommands);
-	assertFalse("c2 has a field named f5.", c2.containsField("f5"));
-	assertTrue("c2 does not have a field named newf5.", c2.containsField("newf5"));
-	assertEquals("c1 made more or less than four fields.", 4, c1.getFields().size());
-	assertEquals("c2 made more or less than two fields.", 2, c2.getFields().size());
-	renameFieldCommands[3] = "newf1";
-	renameFieldCommands[4] = "newf3";
-	control.evaluateCommand(renameFieldCommands);
-	assertFalse("c2 has a field named newf1.", c2.containsField("newf1"));
-	assertTrue("c2 does not have a field named newf3.", c2.containsField("newf3"));
-	assertTrue("c1 does not have a field named newf3.", c1.containsField("newf3"));
-	assertEquals("c1 made more or less than four fields.", 4, c1.getFields().size());
-	assertEquals("c2 made more or less than two fields.", 2, c2.getFields().size());
-	
-	// Rename to an existing field
-	renameFieldCommands[3] = "newf3";
-	renameFieldCommands[4] = "newf5";
-	control.evaluateCommand(renameFieldCommands);
-	assertTrue("c2 does not have a field named newf5.", c2.containsField("newf5"));
-	assertTrue("c2 does not have a field named newf3.", c2.containsField("newf3"));
-	assertTrue("c1 does not have a field named newf3.", c1.containsField("newf3"));
-	assertEquals("c1 made more or less than four fields.", 4, c1.getFields().size());
-	assertEquals("c2 made more or less than two fields.", 2, c2.getFields().size());
-	
-    }
-    
-    @Test
-    public void deleteFieldTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] createFieldCommands = {"create", "field", "c1", "public", "int", "f1"};
-	control.evaluateCommand(createFieldCommands);
-	
-	// Delete one field
-	String[] deleteFieldCommands = {"delete", "field", "c1", "f1"};
-	control.evaluateCommand(deleteFieldCommands);
-	assertFalse("CLI still has a field; Should not have any more fields.", c1.containsField("f1"));
-	assertEquals("c1 has more or less than zero fields.", 0, c1.getFields().size());
-	
-	// Delete multiple fields
-	createFieldCommands[5] = "f2";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f3";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f4";
-	control.evaluateCommand(createFieldCommands);
-	deleteFieldCommands[3] = "f4";
-	control.evaluateCommand(deleteFieldCommands);
-	deleteFieldCommands[3] = "f3";
-	control.evaluateCommand(deleteFieldCommands);
-	deleteFieldCommands[3] = "f2";
-	control.evaluateCommand(deleteFieldCommands);
-	assertFalse("CLI still has a field; Should not have any more fields.", c1.containsField("f2"));
-	assertFalse("CLI still has a field; Should not have any more fields.", c1.containsField("f3"));
-	assertFalse("CLI still has a field; Should not have any more fields.", c1.containsField("f4"));
-	assertEquals("c1 has more or less than zero fields.", 0, c1.getFields().size());
-	
-	// Delete field with wrong length of commands
-	createFieldCommands[5] = "f5";
-	control.evaluateCommand(createFieldCommands);
-	String[] invalidLengthCommand = {"delete", "field", "c1", "f5", "tooLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("CLI deleted field f5; Shouldn't have since parameters are invalid.", c1.containsField("f5"));
-	assertEquals("CLI deleted a field; Shouldn't have since too many arguments.", 1, c1.getFields().size());
-	
-	// Delete field with non-existing class
-	deleteFieldCommands[2] = "c3";
-	deleteFieldCommands[3] = "f5";
-	control.evaluateCommand(deleteFieldCommands);
-	assertTrue("CLI deleted a field in a class that doesn't exist; Should not have", c1.containsField("f5"));
-	assertEquals("c1 has more or less than 1 fields.", 1, c1.getFields().size());
-	
-	// Delete field with non-existing field
-	deleteFieldCommands[2] = "c1";
-	deleteFieldCommands[3] = "f6";
-	control.evaluateCommand(deleteFieldCommands);
-	assertFalse("CLI broke; Should not have", c1.containsField("f6"));
-	assertTrue("CLI deleted the wrong field when the inserted field doesn't exist", c1.containsField("f5"));
-	assertEquals("c1 has more or less than 1 field", 1, c1.getFields().size());	
-	
-	// Delete field in separate classes
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	createFieldCommands[2] = "c2";
-	createFieldCommands[5] = "f5";
-	control.evaluateCommand(createFieldCommands);
-	createFieldCommands[5] = "f6";
-	control.evaluateCommand(createFieldCommands);
-	deleteFieldCommands[2] = "c2";
-	deleteFieldCommands[3] = "f5";
-	control.evaluateCommand(deleteFieldCommands);
-	assertFalse("CLI did not delete a field; Should have deleted f5 in c2.", c2.containsField("f5"));
-	assertTrue("CLI deleted a field; Should have deleted f5 in c1.", c1.containsField("f5"));
-	assertEquals("c1 made more or less than one field.", 1, c1.getFields().size());
-	assertEquals("c2 made more or less than one field.", 1, c2.getFields().size());
-	
-    }
-    
-    //**********************************************************************************************//
-    //**                                METHOD TESTS                                              **//
-    //**                         (CREATE, RENAME, DELETE)                                         **//
-    //**********************************************************************************************//
-    
-    @Test
-    public void createMethodTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	
-	// Create one method
-	String[] methodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI did not make a method; Should be one method called m1.", c1.containsMethod("m1"));
-	assertEquals("c1 made more or less than one method.", 1, c1.getMethods().size());
-	
-	// Create multiple methods
-	methodCommands[5] = "m2";
-	control.evaluateCommand(methodCommands);
-	methodCommands[5] = "m3";
-	control.evaluateCommand(methodCommands);
-	methodCommands[5] = "m4";
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI did not make a method; Should be one method called m2.", c1.containsMethod("m2"));
-	assertTrue("CLI did not make a method; Should be one method called m3.", c1.containsMethod("m3"));
-	assertTrue("CLI did not make a method; Should be one method called m4.", c1.containsMethod("m4"));
-	assertEquals("c1 has more or less than 4 methods", 4, c1.getMethods().size());
-	
-	// Create method with wrong length of commands
-	String[] invalidLengthCommand = {"create", "method", "c1", "public", "int", "tooLong", "badLength"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertFalse("CLI made a method; Shouldn't have since too many arguments.", c1.containsMethod("badLength"));
-	assertFalse("CLI made a method; Shouldn't have since too many arguments.", c1.containsMethod("tooLong"));
-	assertEquals("CLI made a method; Shouldn't have since too many arguments.", 4, c1.getMethods().size());
-	
-	// Create method with non-existing class
-	methodCommands[2] = "c3";
-	methodCommands[5] = "m6";
-	control.evaluateCommand(methodCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertFalse("CLI made a method in a class that doesn't exist; Should not have", c1.containsMethod("m6"));
-	assertEquals("c1 has more or less than 4 methods.", 4, c1.getMethods().size());
-	
-	// Create method with same name
-	methodCommands[5] = "m1";
-	control.evaluateCommand(methodCommands);
-	assertEquals("CLI made a method; Method name already existed.", 4, c1.getMethods().size());
-	
-	// Create method in separate classes
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	methodCommands[2] = "c2";
-	methodCommands[5] = "m5";
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI did not make a method; Should be one method called m5.", c2.containsMethod("m5"));
-	assertEquals("c2 has more or less than one method.", 1, c2.getMethods().size());
-	
-	// Can create the same method name in another class
-	methodCommands[5] = "m1";
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI did not make a method; Should be one method called m1.", c2.containsMethod("m1"));
-	assertTrue("CLI deleted method m5; m5 should still exist.", c2.containsMethod("m5"));
-	assertEquals("c2 has more or less than two methods.", 2, c2.getMethods().size());
-
-	// Can not create a method with same name, but different type
-	methodCommands[4] = "int";
-	methodCommands[5] = "m5";
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI broke", c2.containsMethod("m5"));
-	assertEquals("c2 has more or less than two methods.", 2, c2.getMethods().size());
-    }
-
-    @Test
-    public void renameMethodTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] createMethodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(createMethodCommands);
-	
-	// Rename one method
-	String[] renameMethodCommands = {"rename", "method", "c1", "m1", "newm1"};
-	control.evaluateCommand(renameMethodCommands);
-	assertFalse("c1 still has a m1; Should not have any more m1.", c1.containsMethod("m1"));
-	assertTrue("c1 still has a m1; Should not have any more m1.", c1.containsMethod("newm1"));
-	assertEquals("c1 has more or less than one methods.", 1, c1.getMethods().size());
-	
-	// Rename multiple methods
-	createMethodCommands[5] = "m2";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m3";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m4";
-	control.evaluateCommand(createMethodCommands);
-	renameMethodCommands[3] = "m4";
-	renameMethodCommands[4] = "newm4";
-	control.evaluateCommand(renameMethodCommands);
-	renameMethodCommands[3] = "m3";
-	renameMethodCommands[4] = "newm3";
-	control.evaluateCommand(renameMethodCommands);
-	renameMethodCommands[3] = "m2";
-	renameMethodCommands[4] = "newm2";
-	control.evaluateCommand(renameMethodCommands);
-	assertFalse("c1 still has method m2; Should not have it.", c1.containsMethod("m2"));
-	assertFalse("c1 still has method m3; Should not have it.", c1.containsMethod("m3"));
-	assertFalse("c1 still has method m4; Should not have it.", c1.containsMethod("m4"));
-	assertTrue("c1 doesn't have method newm2; Should have it.", c1.containsMethod("newm2"));
-	assertTrue("c1 doesn't have method newm3; Should have it.", c1.containsMethod("newm3"));
-	assertTrue("c1 doesn't have method newm4; Should have it.", c1.containsMethod("newm4"));
-	assertEquals("c1 has more or less than four methods.", 4, c1.getMethods().size());
-	
-	// Rename method with wrong length of commands
-	String[] invalidLengthCommand = {"rename", "method", "c1", "newm1", "tooLong", "reallyLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("c1 no longer has newm1; Should still have it after an error.", c1.containsMethod("newm1"));
-	assertFalse("c1 renamed method newm1; Shouldn't have since parameters are invalid.", c1.containsMethod("tooLong"));
-	assertFalse("c1 renamed method newm1; Shouldn't have since parameters are invalid.", c1.containsMethod("reallyLong"));
-	assertEquals("c1 has more or less than four methods.", 4, c1.getMethods().size());
-	
-	// Rename method with non-existing class
-	renameMethodCommands[2] = "c3";
-	renameMethodCommands[3] = "newm1";
-	renameMethodCommands[4] = "wrongm1";
-	control.evaluateCommand(renameMethodCommands);
-	assertTrue("c1 no longer has newm1; Should still have it after an error.", c1.containsMethod("newm1"));
-	assertFalse("c1 contains wrongm1; Shouldn't have since class c3 doesn't exist.", c1.containsMethod("wrongm1"));
-	assertEquals("CLI has more or less than one classes.", 1, model.getEntities().size());
-	assertEquals("c1 has more or less than four methods.", 4, c1.getMethods().size());
-	
-	// Rename method with non-existing method
-	renameMethodCommands[2] = "c1";
-	renameMethodCommands[3] = "m5";
-	control.evaluateCommand(renameMethodCommands);
-	assertFalse("CLI broke; Should not have", c1.containsMethod("m5"));
-	assertFalse("c1 contains wrongm1; Should not since m5 doesn't exist", c1.containsMethod("wrongm1"));
-	assertEquals("c1 has more or less than 4 method", 4, c1.getMethods().size());	
-
-	// Rename method in separate classes (with different or same name methods)
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	createMethodCommands[2] = "c2";
-	createMethodCommands[5] = "newm1";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m5";
-	control.evaluateCommand(createMethodCommands);
-	renameMethodCommands[2] = "c2";
-	renameMethodCommands[3] = "m5";
-	renameMethodCommands[4] = "newm5";
-	control.evaluateCommand(renameMethodCommands);
-	
-	assertFalse("c2 has a method named m5.", c2.containsMethod("m5"));
-	assertTrue("c2 does not have a method named newm5.", c2.containsMethod("newm5"));
-	assertEquals("c1 made more or less than four classes.", 4, c1.getMethods().size());
-	assertEquals("c2 made more or less than two classes.", 2, c2.getMethods().size());
-	renameMethodCommands[3] = "newm1";
-	renameMethodCommands[4] = "newm3";
-	control.evaluateCommand(renameMethodCommands);
-	assertFalse("c2 has a method named newm1.", c2.containsMethod("newm1"));
-	assertTrue("c2 does not have a method named newm3.", c2.containsMethod("newm3"));
-	assertTrue("c1 does not have a method named newm3.", c1.containsMethod("newm3"));
-	assertEquals("c1 made more or less than four classes.", 4, c1.getMethods().size());
-	assertEquals("c2 made more or less than two classes.", 2, c2.getMethods().size());
-	
-	// Rename to an existing method
-	renameMethodCommands[3] = "newm3";
-	renameMethodCommands[4] = "newm5";
-	control.evaluateCommand(renameMethodCommands);
-	assertTrue("c2 does not have a method named newm5.", c2.containsMethod("newm5"));
-	assertTrue("c2 does not have a method named newm3.", c2.containsMethod("newm3"));
-	assertTrue("c1 does not have a method named newm3.", c1.containsMethod("newm3"));
-	assertEquals("c1 made more or less than four classes.", 4, c1.getMethods().size());
-	assertEquals("c2 made more or less than two classes.", 2, c2.getMethods().size());
-	
-    }
-    
-    @Test
-    public void deleteMethodTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] createMethodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(createMethodCommands);
-	
-	// Delete one method
-	String[] deleteMethodCommands = {"delete", "method", "c1", "m1"};
-	control.evaluateCommand(deleteMethodCommands);
-	assertFalse("CLI still has a method; Should not have any more methods.", c1.containsMethod("m1"));
-	assertEquals("c1 has more or less than zero methods.", 0, c1.getMethods().size());
-	
-	// Delete multiple methods
-	createMethodCommands[5] = "m2";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m3";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m4";
-	control.evaluateCommand(createMethodCommands);
-	deleteMethodCommands[3] = "m4";
-	control.evaluateCommand(deleteMethodCommands);
-	deleteMethodCommands[3] = "m3";
-	control.evaluateCommand(deleteMethodCommands);
-	deleteMethodCommands[3] = "m2";
-	control.evaluateCommand(deleteMethodCommands);
-	assertFalse("CLI still has a method; Should not have any more methods.", c1.containsMethod("m2"));
-	assertFalse("CLI still has a method; Should not have any more methods.", c1.containsMethod("m3"));
-	assertFalse("CLI still has a method; Should not have any more methods.", c1.containsMethod("m4"));
-	assertEquals("c1 has more or less than zero methods.", 0, c1.getMethods().size());
-	
-	// Delete method with wrong length of commands
-	createMethodCommands[5] = "m5";
-	control.evaluateCommand(createMethodCommands);
-	String[] invalidLengthCommand = {"delete", "method", "c1", "m5", "tooLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("CLI deleted method m5; Shouldn't have since parameters are invalid.", c1.containsMethod("m5"));
-	assertEquals("CLI deleted a method; Shouldn't have since too many arguments.", 1, c1.getMethods().size());
-	
-	// Delete method with non-existing class
-	deleteMethodCommands[2] = "c3";
-	deleteMethodCommands[3] = "m5";
-	control.evaluateCommand(deleteMethodCommands);
-	assertTrue("CLI deleted a method in a class that doesn't exist; Should not have", c1.containsMethod("m5"));
-	assertEquals("c1 has more or less than 1 methods.", 1, c1.getMethods().size());
-	
-	// Delete method with non-existing method
-	deleteMethodCommands[2] = "c1";
-	deleteMethodCommands[3] = "m6";
-	control.evaluateCommand(deleteMethodCommands);
-	assertFalse("CLI broke; Should not have", c1.containsMethod("m6"));
-	assertTrue("CLI deleted the wrong method when the inserted method doesn't exist", c1.containsMethod("m5"));
-	assertEquals("c1 has more or less than 1 method", 1, c1.getMethods().size());	
-	
-	// Delete method in separate classes
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	createMethodCommands[2] = "c2";
-	createMethodCommands[5] = "m5";
-	control.evaluateCommand(createMethodCommands);
-	createMethodCommands[5] = "m6";
-	control.evaluateCommand(createMethodCommands);
-	deleteMethodCommands[2] = "c2";
-	deleteMethodCommands[3] = "m5";
-	control.evaluateCommand(deleteMethodCommands);
-	assertFalse("CLI did not delete a method; Should have deleted m5 in c2.", c2.containsMethod("m5"));
-	assertTrue("CLI deleted a method; Should have deleted m5 in c1.", c1.containsMethod("m5"));
-	assertEquals("c1 made more or less than one class.", 1, c1.getMethods().size());
-	assertEquals("c2 made more or less than one class.", 1, c2.getMethods().size());
-	
-    }
-    
-    //**********************************************************************************************//
-    //**                              PARAMETER TESTS                                             **//
-    //**                         (CREATE, RENAME, DELETE)                                         **//
-    //**********************************************************************************************//
-    
-    @Test
-    public void createParameterTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] methodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(methodCommands);
-	Method m1 = c1.getMethods().get(0);
-	
-	// Create one parameter
-	String[] parameterCommands = {"create", "parameter", "c1", "m1", "String", "p1"};
-	control.evaluateCommand(parameterCommands);
-	assertTrue("CLI did not make a parameter; Should be one parameter called p1.", m1.containsParameter("p1"));
-	assertEquals("c1 made more or less than one methods.", 1, c1.getMethods().size());
-	assertEquals("m1 made more or less than one parameter.", 1, m1.getParameters().size());
-	
-	// Create multiple parameters
-	parameterCommands[5] = "p2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p3";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p4";
-	control.evaluateCommand(parameterCommands);
-	assertTrue("CLI did not make a parameter; Should be one parameter called p2.", m1.containsParameter("p2"));
-	assertTrue("CLI did not make a parameter; Should be one parameter called p3.", m1.containsParameter("p3"));
-	assertTrue("CLI did not make a parameter; Should be one parameter called p4.", m1.containsParameter("p4"));
-	assertEquals("m1 has more or less than 4 parameters", 4, m1.getParameters().size());
-	
-	// Create parameter with wrong length of commands
-	String[] invalidLengthCommand = {"create", "parameter", "c1", "m1", "int", "p5", "tooLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertFalse("CLI made a method; Shouldn't have since too many arguments.", m1.containsParameter("p5"));
-	assertFalse("CLI made a method; Shouldn't have since too many arguments.", m1.containsParameter("tooLong"));
-	assertEquals("CLI made a method; Shouldn't have since too many arguments.", 4, m1.getParameters().size());
-	
-	// Create parameter with non-existing class
-	parameterCommands[2] = "c3";
-	parameterCommands[5] = "p5";
-	control.evaluateCommand(parameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertFalse("CLI made a parameter in a class that doesn't exist; Should not have", m1.containsParameter("p5"));
-	assertEquals("m1 has more or less than 4 methods.", 4, m1.getParameters().size());
-	
-	// Create parameter with non-existing method
-	parameterCommands[2] = "c1";
-	parameterCommands[3] = "m2";
-	control.evaluateCommand(parameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertEquals("CLI has more or less than one class", 1, c1.getMethods().size());
-	assertTrue("c1 no longer contains method m1", c1.containsMethod("m1"));
-	assertFalse("CLI made a parameter in a method that doesn't exist; Should not have", m1.containsParameter("p5"));
-	assertEquals("m1 has more or less than 4 methods.", 4, m1.getParameters().size());
-	
-	// Create parameter with same name
-	parameterCommands[3] = "m1";
-	parameterCommands[5] = "p1";
-	control.evaluateCommand(methodCommands);
-	assertTrue("CLI broke; should not have", m1.containsParameter("p1"));
-	assertEquals("CLI made a parameter; Parameter name already existed.", 4, m1.getParameters().size());
-	
-	// Create Parameter in separate classes/methods (with same and/or different names)
-	methodCommands[5] = "m2";
-	control.evaluateCommand(methodCommands);
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	methodCommands[2] = "c2";
-	methodCommands[5] = "m3";
-	control.evaluateCommand(methodCommands);
-	parameterCommands[3] = "m2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[2] = "c2";
-	parameterCommands[3] = "m3";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "newp1";
-	control.evaluateCommand(parameterCommands);
-	assertTrue("CLI broke; should not have.", m1.containsParameter("p1"));
-	assertEquals("m1 has more or less than four parameters.", 4, m1.getParameters().size());
-	assertTrue("CLI did not make a parameter in m2; Should be one parameter called p1 in m2.", c1.getMethods().get(1).containsParameter("p1"));
-	assertEquals("m2 has more or less than one parameter.", 1, c1.getMethods().get(1).getParameters().size());
-	assertTrue("CLI did not make a parameter in m3; Should be one parameter called p1 in m3.", c2.getMethods().get(0).containsParameter("p1"));
-	assertTrue("CLI did not make a parameter in m3; Should be one parameter called p1 in m3.", c2.getMethods().get(0).containsParameter("newp1"));
-	assertEquals("m3 has more or less than two parameter.", 2, c2.getMethods().get(0).getParameters().size());
-	
-	// Create a parameter with same name, but different type
-	parameterCommands[4] = "int";
-	parameterCommands[5] = "newp1";
-	control.evaluateCommand(parameterCommands);
-	assertTrue("CLI did not break", c2.getMethods().get(0).containsParameter("newp1"));
-	assertEquals("m3 has more or less than two parameter.", 2, c2.getMethods().get(0).getParameters().size());
-	
-    }
-    
-    @Test
-    public void renameParameterTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] methodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(methodCommands);
-	Method m1 = c1.getMethods().get(0);
-	String[] parameterCommands = {"create", "parameter", "c1", "m1", "int", "p1"};
-	control.evaluateCommand(parameterCommands);
-	
-	// Rename one parameter
-	String[] renameParameterCommands = {"rename", "parameter", "c1", "m1", "p1", "newp1"};
-	control.evaluateCommand(renameParameterCommands);
-	assertTrue("CLI did not rename a parameter; Should be one parameter called newp1.", m1.containsParameter("newp1"));
-	assertFalse("CLI did not get rid of p1.", m1.containsParameter("p1"));
-	assertEquals("c1 made more or less than one methods.", 1, c1.getMethods().size());
-	assertEquals("m1 made more or less than one parameter.", 1, m1.getParameters().size());
-	
-	// Rename back to original name
-	renameParameterCommands[4] = "newp1";
-	renameParameterCommands[5] = "p1";
-	control.evaluateCommand(renameParameterCommands);
-	assertFalse("CLI did not rename a parameter; Should be one parameter called p1.", m1.containsParameter("newp1"));
-	assertTrue("CLI did not get change to p1.", m1.containsParameter("p1"));
-	assertEquals("c1 made more or less than one methods.", 1, c1.getMethods().size());
-	assertEquals("m1 made more or less than one parameter.", 1, m1.getParameters().size());
-	
-	// Rename multiple parameters
-	parameterCommands[5] = "p2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p3";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p4";
-	control.evaluateCommand(parameterCommands);
-	renameParameterCommands[4] = "p2";
-	renameParameterCommands[5] = "newp2";
-	control.evaluateCommand(renameParameterCommands);
-	renameParameterCommands[4] = "p3";
-	renameParameterCommands[5] = "newp3";
-	control.evaluateCommand(renameParameterCommands);
-	renameParameterCommands[4] = "p4";
-	renameParameterCommands[5] = "newp4";
-	control.evaluateCommand(renameParameterCommands);
-	assertFalse("CLI did not get rid of p2.", m1.containsParameter("p2"));
-	assertFalse("CLI did not get rid of p3.", m1.containsParameter("p3"));
-	assertFalse("CLI did not get rid of p4.", m1.containsParameter("p4"));
-	assertTrue("CLI did not make a parameter; Should be one parameter called newp2.", m1.containsParameter("newp2"));
-	assertTrue("CLI did not make a parameter; Should be one parameter called newp3.", m1.containsParameter("newp3"));
-	assertTrue("CLI did not make a parameter; Should be one parameter called newp4.", m1.containsParameter("newp4"));
-	assertEquals("m1 has more or less than 4 parameters", 4, m1.getParameters().size());
-	
-	// Rename parameter with wrong length of commands
-	String[] invalidLengthCommand = {"rename", "parameter", "c1", "m1", "newp4", "tooLong", "reallyLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("CLI deleted a parameter; Shouldn't have since too many arguments.", m1.containsParameter("newp4"));
-	assertFalse("CLI made a parameter; Shouldn't have since too many arguments.", m1.containsParameter("tooLong"));
-	assertFalse("CLI made a paremeter; Shouldn't have since too many arguments.", m1.containsParameter("reallyLong"));
-	assertEquals("CLI made a parameter; Shouldn't have since too many arguments.", 4, m1.getParameters().size());
-	
-	// Rename parameter with non-existing class
-	renameParameterCommands[2] = "c3";
-	renameParameterCommands[4] = "newp4";
-	renameParameterCommands[5] = "p5";
-	control.evaluateCommand(renameParameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertTrue("m1 no longer has new p4.", m1.containsParameter("newp4"));
-	assertFalse("CLI made a parameter in a class that doesn't exist; Should not have", m1.containsParameter("p5"));
-	assertEquals("m1 has more or less than 4 methods.", 4, m1.getParameters().size());
-	
-	// Rename parameter with non-existing method
-	renameParameterCommands[2] = "c1";
-	renameParameterCommands[3] = "m2";
-	control.evaluateCommand(renameParameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertEquals("CLI has more or less than one method", 1, c1.getMethods().size());
-	assertTrue("c1 no longer contains method m1", c1.containsMethod("m1"));
-	assertTrue("m1 no longer has new p4.", m1.containsParameter("newp4"));
-	assertFalse("CLI made a parameter in a method that doesn't exist; Should not have", m1.containsParameter("p5"));
-	assertEquals("m1 has more or less than 4 methods.", 4, m1.getParameters().size());
-	
-	// Rename parameter with non-existing parameter
-	renameParameterCommands[3] = "m1";
-	renameParameterCommands[4] = "p5";
-	renameParameterCommands[5] = "newp5";
-	control.evaluateCommand(renameParameterCommands);
-	assertFalse("CLI broke; Should not have", m1.containsParameter("p5"));
-	assertFalse("CLI broke; Should not have", m1.containsParameter("newp5"));
-	assertEquals("m1 has more or less than 4 parameters", 4, m1.getParameters().size());
-	
-	// Rename parameter with same name
-	renameParameterCommands[3] = "m1";
-	renameParameterCommands[4] = "newp2";
-	renameParameterCommands[5] = "p1";
-	control.evaluateCommand(renameParameterCommands);
-	assertTrue("CLI broke; should not have", m1.containsParameter("p1"));
-	assertTrue("CLI got rid of newp2; should not have", m1.containsParameter("newp2"));
-	assertEquals("CLI made a parameter; Parameter name already existed.", 4, m1.getParameters().size());
-	
-	// Rename Parameter in separate classes/methods (with same and/or different names)
-	methodCommands[5] = "m2";
-	control.evaluateCommand(methodCommands);
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	methodCommands[2] = "c2";
-	methodCommands[5] = "m3";
-	control.evaluateCommand(methodCommands);
-	parameterCommands[3] = "m2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[2] = "c2";
-	parameterCommands[3] = "m3";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "newp1";
-	control.evaluateCommand(parameterCommands);
-	renameParameterCommands[3] = "m2";
-	renameParameterCommands[4] = "p4";
-	renameParameterCommands[5] = "newp4";
-	control.evaluateCommand(renameParameterCommands);
-	renameParameterCommands[2] = "c2";
-	renameParameterCommands[3] = "m3";
-	renameParameterCommands[4] = "p4";
-	renameParameterCommands[5] = "newp4";
-	control.evaluateCommand(renameParameterCommands);
-	renameParameterCommands[4] = "newp1";
-	renameParameterCommands[5] = "p5";
-	control.evaluateCommand(renameParameterCommands);	
-	assertTrue("CLI broke; should not have.", m1.containsParameter("newp4"));
-	assertEquals("m1 has more or less than four parameters.", 4, m1.getParameters().size());
-	assertFalse("CLI did not get rid of a parameter in m2; Should be one parameter called newp4 in m2.", c1.getMethods().get(1).containsParameter("p4"));
-	assertTrue("CLI did not rename a parameter in m2; Should be one parameter called newp4 in m2.", c1.getMethods().get(1).containsParameter("newp4"));
-	assertEquals("m2 has more or less than one parameter.", 1, c1.getMethods().get(1).getParameters().size());
-	assertFalse("CLI did not get rid of a parameter in m3; Should be a parameter called newp4 in m3.", c2.getMethods().get(0).containsParameter("p4"));
-	assertTrue("CLI did not rename a parameter in m3; Should be a parameter called newp4 in m3.", c2.getMethods().get(0).containsParameter("newp4"));
-	assertFalse("CLI did not get rid of a parameter in m3; Should be a parameter called np5 in m3.", c2.getMethods().get(0).containsParameter("newp5"));
-	assertTrue("CLI did not rename a parameter in m3; Should be a parameter called p5 in m3.", c2.getMethods().get(0).containsParameter("p5"));
-	assertEquals("m3 has more or less than two parameter.", 2, c2.getMethods().get(0).getParameters().size());
-	
-    }
-    
-    @Test
-    public void deleteParameterTest() {
-	String[] classCommands = {"create", "class", "c1"};
-	control.evaluateCommand(classCommands);
-	Entity c1 = model.getEntities().get(0);
-	String[] methodCommands = {"create", "method", "c1", "public", "int", "m1"};
-	control.evaluateCommand(methodCommands);
-	Method m1 = c1.getMethods().get(0);
-	String[] parameterCommands = {"create", "parameter", "c1", "m1", "int", "p1"};
-	control.evaluateCommand(parameterCommands);
-	
-	// Delete one parameter
-	String[] deleteParameterCommands = {"delete", "parameter", "c1", "m1", "p1"};
-	control.evaluateCommand(deleteParameterCommands);
-	assertFalse("CLI did not get rid of p1.", m1.containsParameter("p1"));
-	assertEquals("c1 made more or less than one methods.", 1, c1.getMethods().size());
-	assertEquals("m1 made more or less than zero parameter.", 0, m1.getParameters().size());
-	
-	// Delete multiple parameters
-	parameterCommands[5] = "p1";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "p3";
-	control.evaluateCommand(parameterCommands);
-	deleteParameterCommands[4] = "p3";
-	control.evaluateCommand(deleteParameterCommands);
-	deleteParameterCommands[4] = "p2";
-	control.evaluateCommand(deleteParameterCommands);
-	deleteParameterCommands[4] = "p1";
-	control.evaluateCommand(deleteParameterCommands);
-	assertFalse("CLI did not get rid of p2.", m1.containsParameter("p1"));
-	assertFalse("CLI did not get rid of p3.", m1.containsParameter("p2"));
-	assertFalse("CLI did not get rid of p4.", m1.containsParameter("p3"));
-	assertEquals("m1 has more or less than zero parameters.", 0, m1.getParameters().size());
-	
-	// Delete parameter with wrong length of commands
-	parameterCommands[5] = "p1";
-	control.evaluateCommand(parameterCommands);
-	String[] invalidLengthCommand = {"delete", "parameter", "c1", "m1", "p1", "tooLong"};
-	control.evaluateCommand(invalidLengthCommand);
-	assertTrue("CLI deleted a parameter; Shouldn't have since too many arguments.", m1.containsParameter("p1"));
-	assertFalse("CLI made a parameter; Shouldn't have since too many arguments.", m1.containsParameter("tooLong"));
-	assertEquals("m1 has more or less than one parameter.", 1, m1.getParameters().size());
-	
-	// Delete parameter with non-existing class
-	deleteParameterCommands[2] = "c3";
-	control.evaluateCommand(deleteParameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertTrue("m1 no longer has a p1.", m1.containsParameter("p1"));
-	assertEquals("m1 has more or less than 1 parameters.", 1, m1.getParameters().size());
-	
-	// Delete parameter with non-existing method
-	deleteParameterCommands[2] = "c1";
-	deleteParameterCommands[3] = "m2";
-	control.evaluateCommand(deleteParameterCommands);
-	assertEquals("CLI has more or less than one class", 1, model.getEntities().size());
-	assertEquals("CLI has more or less than one method", 1, c1.getMethods().size());
-	assertTrue("c1 no longer contains method m1", c1.containsMethod("m1"));
-	assertTrue("m1 no longer has new p4.", m1.containsParameter("p1"));
-	assertEquals("m1 has more or less than one parameter.", 1, m1.getParameters().size());
-	
-	// Delete parameter with non-existing parameter
-	deleteParameterCommands[3] = "m1";
-	deleteParameterCommands[4] = "p2";
-	control.evaluateCommand(deleteParameterCommands);
-	assertFalse("CLI broke; Should not have", m1.containsParameter("p2"));
-	assertEquals("m1 has more or less than one parameters", 1, m1.getParameters().size());
-	
-	// Delete parameter in separate classes/methods (with same and/or different names)
-	methodCommands[5] = "m2";
-	control.evaluateCommand(methodCommands);
-	classCommands[2] = "c2";
-	control.evaluateCommand(classCommands);
-	Entity c2 = model.getEntities().get(1);
-	methodCommands[2] = "c2";
-	methodCommands[5] = "m3";
-	control.evaluateCommand(methodCommands);
-	parameterCommands[3] = "m2";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[2] = "c2";
-	parameterCommands[3] = "m3";
-	control.evaluateCommand(parameterCommands);
-	parameterCommands[5] = "newp1";
-	control.evaluateCommand(parameterCommands);
-	deleteParameterCommands[2] = "c2";
-	deleteParameterCommands[3] = "m3";
-	deleteParameterCommands[4] = "newp1";
-	control.evaluateCommand(deleteParameterCommands);
-	deleteParameterCommands[4] = "p1";
-	control.evaluateCommand(deleteParameterCommands);
-	deleteParameterCommands[2] = "c1";
-	deleteParameterCommands[3] = "m2";
-	control.evaluateCommand(deleteParameterCommands);
-	deleteParameterCommands[3] = "m1";
-	control.evaluateCommand(deleteParameterCommands);
-	assertFalse("CLI broke; should not have.", m1.containsParameter("p1"));
-	assertEquals("m1 has more or less than zero parameters.", 0, m1.getParameters().size());
-	assertFalse("CLI did not get rid of a parameter in m2; Should be no more parameters in m2.", c1.getMethods().get(1).containsParameter("p1"));
-	assertEquals("m2 has more or less than zero parameters.", 0, c1.getMethods().get(1).getParameters().size());
-	assertFalse("CLI did not get rid of a parameter in m3; Should be no more parameters in m3.", c2.getMethods().get(0).containsParameter("p1"));
-	assertFalse("CLI did not get rid of a parameter in m3; Should be no more parameters in m3.", c2.getMethods().get(0).containsParameter("newp1"));
-	assertEquals("m3 has more or less than zero parameter.", 0, c2.getMethods().get(0).getParameters().size());
-    	
-    }
-    
-    //**********************************************************************************************//
-    //**                             RELATIONSHIP TESTS                                           **//
-    //**                              (CREATE, DELETE)                                            **//
-    //**********************************************************************************************//
-
-    @Test
-    public void createRelationshipTest() {
-        //create classes to have relationships
-        String[] classes = {"create", "class", "c1"};
-        control.evaluateCommand(classes);
-        classes[2] = "c2";
-        control.evaluateCommand(classes);
-        classes[2] = "c3";
-        control.evaluateCommand(classes);
-
-        //attempting to create relationships to non-existent classes
-        String[] wrongRels = {"create", "relationship", "aggregation", "fake1", "fake2"};
-        control.evaluateCommand(wrongRels);
-        assertTrue("Relationship was created with fake classes", model.getRelationships().isEmpty());
-
-        //creating relationship with invalid type
-        wrongRels[2] = "fake";
-        wrongRels[3] = "c1";
-        wrongRels[4] = "c2";
-        control.evaluateCommand(wrongRels);
-        assertTrue("Relationship was created with invalid type", model.getRelationships().isEmpty());
-
-        //create relationships of each type
-        String[] relationships = {"create", "relationship", "aggregation", "c1", "c2"};
-        control.evaluateCommand(relationships);
-        assertTrue("Aggregation relationship between c1 and c2 not found", model.containsRelationship(Type.AGGREGATION, "c1", "c2"));
-
-        relationships[2] = "composition";
-        relationships[3] = "c2";
-        relationships[4] = "c3";
-
-        control.evaluateCommand(relationships);
-        assertTrue("Composition relationship between c2 and c3 not found", model.containsRelationship(Type.COMPOSITION, "c2", "c3"));
-
-        relationships[2] = "inheritance";
-        relationships[3] = "c1";
-
-        control.evaluateCommand(relationships);
-        assertTrue("Inheritance relationship between c1 and c3 not found", model.containsRelationship(Type.INHERITANCE, "c1", "c3"));
-        
-        // Create relationship wrong length
-        resetStreams();
-        System.out.println("\nERROR: "
-        	+ "Error in parsing command. Proper command usage is: \n"
-        	+ "  create relationship <type> <class name1> <class name2>\n");
-	String expected = out.toString();
-	resetStreams();
-	String[] createErr = { "create", "relationship", "WRONG" };
-	control.evaluateCommand(createErr);
-	assertEquals("relationship error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-        
-        //recursive relationship
-        relationships[2] = "realization";
-        relationships[3] = "c2";
-        relationships[4] = "c2";
-
-        control.evaluateCommand(relationships);
-        assertTrue("Recursive realization relationship associated with c2 not found", model.containsRelationship(Type.REALIZATION, "c2", "c2"));
-
-        //attempting to create a relationship that already exists
-        control.evaluateCommand(relationships);
-        assertEquals("Recursive realization c2 relationship created twice", 4, model.getRelationships().size());
-
-        relationships[2] = "composition";
-        assertEquals("New c2 recursive relationship created, even though c2 already has a recursive relationship", 4, model.getRelationships().size());
-    }
-
-    @Test
-    public void deleteRelationshipTest() {
-        //create classes to have relationships
-        String[] classes = {"create", "class", "c1"};
-        control.evaluateCommand(classes);
-        classes[2] = "c2";
-        control.evaluateCommand(classes);
-        classes[2] = "c3";
-        control.evaluateCommand(classes);
-
-        //create relationships (c1--c2, c2--c3, c3--c3)
-        String[] relationships = {"create", "relationship", "aggregation", "c1", "c2"};
-        control.evaluateCommand(relationships);
-
-        relationships[2] = "realization";
-        relationships[3] = "c2";
-        relationships[4] = "c3";
-        control.evaluateCommand(relationships);
-
-        relationships[2] = "inheritance";
-        relationships[3] = "c3";
-        control.evaluateCommand(relationships);
-
-        //deleting non-existent relationships
-        relationships[0] = "delete";
-        relationships[2] = "fakeType";
-        control.evaluateCommand(relationships);
-        assertTrue("c3 recursive relationship deleted despite fake relationship type", model.containsRelationship(Type.INHERITANCE, "c3", "c3"));
-
-        relationships[2] = "realization";
-        control.evaluateCommand(relationships);
-        assertTrue("Relationship c3--c3 deleted despite wrong relationship type given", model.containsRelationship(Type.INHERITANCE, "c3", "c3"));
-
-        relationships[3] = "fake1";
-        relationships[4] = "fake2";
-        control.evaluateCommand(relationships);
-        assertEquals("Relationship deleted despite fake class names", 3, model.getRelationships().size());
-
-        //deleting relationships
-        relationships[2] = "realization";
-        relationships[3] = "c2";
-        relationships[4] = "c3";
-        control.evaluateCommand(relationships);
-        assertFalse("Relationship c2--c3 failed to delete", model.containsRelationship(Type.REALIZATION, "c2", "c3"));
-        assertEquals("Size of relationship list is wrong", 2, model.getRelationships().size());
-
-        relationships[2] = "inheritance";
-        relationships[3] = "c3";
-        control.evaluateCommand(relationships);
-        assertFalse("Relationship c3--c3 failed to delete", model.containsRelationship(Type.INHERITANCE, "c3", "c3"));
-        assertEquals("Size of relationship list is wrong", 1, model.getRelationships().size());
-    }
-
-    //**********************************************************************************************//
-    //**                                MISC. TESTS                                               **//
-    //**                         	  (CLEAR)                                                 **//
-    //**********************************************************************************************//
-	
-    @Test
-    public void clearTest() {
-        // Make a whole bunch of stuff to be cleared.
-        String[] classes = {"create", "class", "c1"};
-        control.evaluateCommand(classes);
-        classes[2] = "c2";
-        control.evaluateCommand(classes);
-
-        assertEquals("CLI did not make the classes; Should be 2.", 2, model.getEntities().size());
-
-        String[] field = {"create", "field", "c1", "public", "int", "f1"};
-        control.evaluateCommand(field);
-        field[5] = "f2";
-        control.evaluateCommand(field);
-        field[2] = "c2";
-        field[5] = "f1";
-        control.evaluateCommand(field);
-
-        assertTrue("CLI did not make 2 fields in c1; C1 should have f1 and f2.", model.getEntities().get(0).containsField("f1"));
-        assertTrue("CLI did not make 2 fields in c1; C1 should have f1 and f2.", model.getEntities().get(0).containsField("f2"));
-        assertTrue("CLI did not make 1 field in c2; C2 should have f1.", model.getEntities().get(1).containsField("f1"));
-
-        String[] method = {"create", "method", "c1", "public", "double", "m1"};
-        control.evaluateCommand(method);
-        method[5] = "m2";
-        control.evaluateCommand(method);
-        method[2] = "c2";
-        method[5] = "m1";
-        control.evaluateCommand(method);
-
-        assertTrue("CLI did not make 2 methods in c1; C1 should have m1 and m2.", model.getEntities().get(0).containsMethod("m1"));
-        assertTrue("CLI did not make 2 methods in c1; C1 should have m1 and m2.", model.getEntities().get(0).containsMethod("m2"));
-        assertTrue("CLI did not make 1 method in c2; C2 should have m1.", model.getEntities().get(1).containsMethod("m1"));
-
-        String[] parameter = {"create", "parameter", "c1", "m1", "char", "p1"};
-        control.evaluateCommand(parameter);
-        parameter[5] = "p2";
-        control.evaluateCommand(parameter);
-        parameter[3] = "m2";
-        parameter[5] = "p1";
-        control.evaluateCommand(parameter);
-        parameter[2] = "c2";
-        parameter[3] = "m1";
-        control.evaluateCommand(parameter);
-
-        assertTrue("CLI did not make 2 parameters in c1 > m1; M1 should have p1 and p2.", model.getEntities().get(0).getMethods().get(0).containsParameter("p1"));
-        assertTrue("CLI did not make 2 parameters in c1 > m1; M1 should have p1 and p2.", model.getEntities().get(0).getMethods().get(0).containsParameter("p2"));
-        assertTrue("CLI did not make 1 parameter in c1 > m2; M2 should have p1.", model.getEntities().get(0).getMethods().get(1).containsParameter("p1"));
-        assertTrue("CLI did not make 1 parameter in c2 > m1; M1 should have p1.", model.getEntities().get(1).getMethods().get(0).containsParameter("p1"));
-
-        String[] relationship = {"create", "relationship", "realization", "c1", "c1"};
-        control.evaluateCommand(relationship);
-        relationship[2] = "aggregation";
-        relationship[4] = "c2";
-        control.evaluateCommand(relationship);
-
-        assertTrue("CLI did not make a recursive relationship. There should be a realization from c1 to c1.", model.containsRelationship(Type.REALIZATION, "c1", "c1"));
-        assertTrue("CLI did not make a normal relationship. There should be an aggregation from c1 to c2.", model.containsRelationship(Type.AGGREGATION, "c1", "c2"));
-
-
-        String[] clear = {"sudo", "clear"};
-        control.evaluateCommand(clear);
-
-        assertTrue("CLI's model is not empty; CLI's model should be empty.", model.empty());
-    }
-    
-    @Test
-    public void helpTest() {
-	String[] help = {"help"};
-	ByteArrayOutputStream out = new ByteArrayOutputStream();
-	ByteArrayOutputStream err = new ByteArrayOutputStream();
-	
-	System.setOut(new PrintStream(out));
-	System.setErr(new PrintStream(err));
-	
-	System.out.println("\n"
-		+ "Here is a list of available commands:\n"
-		+ "  save <name>.json (optional <path>) - Save file to specific path\n"
-		+ "  load <path> - Loads a file at a specific path\n"
-		+ "\n"
-		+ "  create class <name> - create a class with title <name>\n"
-		+ "  create field <class name> <field visibility> <field type> <field name> - create a field in <class name> with visibility <type visibility>, type <field type> titled <field name>\n"
-		+ "  create method <class name> <method visibility> <method type> <method name> - create a method in <class name> with visibility <method visibility>, type <method type> titled <method name>\n"
-		+ "  create relationship <type> <class name1> <class name2> - create a relationship between <class name1> and <class name2> with type <type> (Aggregation, Realization, Composition, Inheritance)\n"
-		+ "  create parameter <class name> <method> <parameter type> <parameter name> - create a parameter in <class name> for <method> with type <parameter type> titled <parameter name>\n"
-		+ "\n"
-		+ "  delete class <name> - delete a class with title <name>\n"
-		+ "  delete field <class name> <field name> - delete field <field name> in class titled <class name>\n"
-		+ "  delete method <class name> <method name> - delete method <method name> in class titled <class name>\n"
-		+ "  delete relationship <type> <class name1> <class name2> - delete a relationship with type <type> (Aggregation, Realization, Composition, Inheritance) between <class name1> and <class name2>\n"
-		+ "  delete parameter <class name> <method name>, <parameter name> - delete a parameter in <class name> for <method name> with  <parameter name>\n"
-		+ "\n"
-		+ "  rename class <name> <newname> - rename class <name> to <new name>\n"
-		+ "  rename field <class name> <field name> <newname> - rename field <field name> to <newname> in class titled <class name>\n"
-		+ "  rename method <class name> <method name> <newname> - rename method <method name> to <newname> in class titled <class name>\n"
-		+ "  rename parameter <class name> <method name> <parameter name> <parameter newname> - rename parameter in <class name> for <method> titled <parameter name> to <parameter newname>\n"
-		+ "\n"
-		+ "  settype field <class name> <field name> <newtype> - set type of field <field name> in <class name> to <type>\n"
-		+ "  settype method <class name> <method name> <newtype> - set type of method <method name> in <class name> to <type>\n"
-		+ "  settype parameter <class name> <method name> <parameter name> <newtype> - set type of parameter <class name> in <method name> titled <parameter name> to <type>\n"
-		+ "\n"
-		+ "  setvis field <class name> <field name> <visibility> - set visibility of field <field name> in <class name> to <visibility>\n"
-		+ "  setvis method <class name> <method name> <visibility> - set visibility of method <method name> in <class name> to <visibility>\n"
-		+ "\n"
-		+ "  list classes - List all existing classes\n"
-		+ "  list relationships - List all existing relationships\n"
-		+ "  list all - List all existing classes and relationships\n"
-		+ "\n"
-		+ "  clear - Clear all classes and relationships\n"
-		+ "  quit - exits the program"
-		+ "\n");
-	String expected = out.toString();
-	out.reset();
-	err.reset();
-        control.evaluateCommand(help);
-	assertEquals("Initial print all does not equal printout", expected, out.toString());
-    }
-
-    @Test
-    public void saveTest() throws IOException, ParseException {
-	// Test relative Path error length 4
-	System.out.println("\nERROR: "
-		+ "Error in parsing command. Proper command usage is: \n"
-		+ "  save <name>.json (optional <path>)\n");
-	String expected = out.toString();
-	resetStreams();
-	String[] saveErr2 = {"save", "testDemoCLI.json", "\\src\\test\\java\\mike", "WRONG"};
-	control.evaluateCommand(saveErr2);
-	assertEquals("save error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	
-	// Test relative Path
-	String[] save = {"save", "\\src\\test\\java\\mike\\testDemoCLI.json"};
-	control.evaluateCommand(save);
-	Path path = Paths.get(System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json");
-	saveWorked(path);
-	
-	// Test absolute Path
-	save[1] = System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json";
-	control.evaluateCommand(save);
-	path = Paths.get(save[1]);
-	saveWorked(path);
-	
-	// Test giving relative directory
-	String[] saveDir = {"save", "testDemoCLI.json", "\\src\\test\\java\\mike" };
-	control.evaluateCommand(saveDir);
-	Path pathDir = Paths.get(System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json");
-	saveWorked(pathDir);
-	
-	// Test giving absolute directory
-	saveDir[2] = System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json";
-	control.evaluateCommand(saveDir);
-	pathDir = Paths.get(System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json");
-	saveWorked(pathDir);
-	
-    }
-    
-    @Test
-    public void loadTest() throws IOException, ParseException, java.text.ParseException {
-	// Test relative Path error length 3
-	System.out.println("\nERROR: "
-		+ "Error in parsing command. Proper command usage is: \n"
-		+ "  load <path>\n");
-	String expected = out.toString();
-	resetStreams();
-	String[] loadErr2 = {"load", "testDemo.json", "WRONG"};
-	control.evaluateCommand(loadErr2);
-	assertEquals("load error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	
-	// Test relative Path error length 2
-	System.out.println("\nERROR: Failed to parse directory. Exiting.");
-	expected = out.toString();
-	resetStreams();
-	String[] loadErr3 = {"load", "\\src\\test\\java\\mike\\testDemo2.json"};
-	control.evaluateCommand(loadErr3);
-	assertEquals("load error message did not appear correctly.", expected, out.toString());
-	
-	// Test relative Path
-	String[] load = {"load", "\\src\\test\\java\\mike\\testDemoCLI.json"};
-	control.evaluateCommand(load);
-	Path path = Paths.get(System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json");
-	loadWorked(path);
-	
-	// Test absolute Path
-	load[1] = System.getProperty("user.dir") + "\\src\\test\\java\\mike\\testDemoCLI.json";
-	control.evaluateCommand(load);
-	path = Paths.get(load[1]);
-	loadWorked(path);
-	
-    }
-    
-    @Test
-    public void createErr() {
-	System.out.println("\nERROR: "
-		+ "Error in parsing command. Proper command usage is: \n"
-		+ "  create class <name>\n"
-		+ "  create field <class name> <field visibility> <field type> <field name>\n"
-		+ "  create method <class name> <method visibility> <method type> <method name>\n"
-		+ "  create relationship <type> <class name1> <class name2>\n"
-		+ "  create parameter <class name> <method> <parameter type> <parameter name>\n");
-	String expected = out.toString();
-	resetStreams();
-	String[] createErr2 = {"create"};
-	control.evaluateCommand(createErr2);
-	assertEquals("load error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	
-	System.out.println("\nERROR: "
-		+ "Invalid visibility type. Valid types are public, private, or protected.\n");
-	expected = out.toString();
-	resetStreams();
-	String[] createErr = { "create", "field", "c1", "wrong", "int", "f1" };
-	control.evaluateCommand(createErr);
-	assertEquals("load error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	
-	System.out.println("\nERROR: "
-		+ "Invalid visibility type. Valid types are public, private, or protected.\n");
-	expected = out.toString();
-	resetStreams();
-	createErr[1] = "method";
-	createErr[5] = "m1";
-	control.evaluateCommand(createErr);
-	assertEquals("load error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	/*
-	System.out.println("\n"
-		+ "Error in parsing command. Proper command usage is: \n"
-		+ "  create class <name>\n"
-		+ "  create field <class name> <field visibility> <field type> <field name>\n"
-		+ "  create method <class name> <method visibility> <method type> <method name>\n"
-		+ "  create relationship <type> <class name1> <class name2>\n"
-		+ "  create parameter <class name> <method> <parameter type> <parameter name>\n");
-	expected = out.toString();
-	resetStreams();
-	String[] createErr3 = { "create", "Error" };
-	cli.evaluateCommand(createErr3);
-	assertEquals("Error message did not appear correctly.", expected, out.toString());
-	resetStreams();
-	*/
-    }
-    
-    private void saveWorked(Path path) throws IOException, ParseException {
-	HelperMethods.save(path, model);
-	File directory = new File(path.toString());
-	Object obj = new JSONParser().parse(new FileReader(directory));
-	
-	String emptyModelFile = "{\"Relationships\":[],\"Classes\":[]}";
-	assertEquals("CLI is null; Should not be null.", emptyModelFile, obj.toString());
-    }
-    
-    private void loadWorked(Path path) throws IOException, ParseException, java.text.ParseException {
-	HelperMethods.save(path, model);
-	Model loadModel = new Model();
-	HelperMethods.load(path, loadModel, null, null);
-	
-	assertEquals("loadModel contains a class.", 0, loadModel.getEntities().size());
-	assertEquals("loadModel contains a relationship.", 0, loadModel.getRelationships().size());
-    }
-    
-    
-    
 }
